Remarks fixes
