+++ src/routes/transactions/components/desktop/TransactionsTable.tsx
@@ -2,7 +2,7 @@ import { createStyles, withStyles, WithStyles } from "@material-ui/core";
 import * as React from "react";
 import Block from "~/components/layout/Block";
 import { background, shadowColor } from "~/theme/variables";
-import { TransactionsTableState, TransactionTableProps } from "../index";
+import { TransactionsTableState, TransactionTableProps } from "../../common";
 import TransactionTableFooter from "../TransactionTableFooter";
 import TransactionRow from "./TransactionRow";
 import TransactionsTableHeader from "./TransactionsTableHeader";
@@ -33,7 +33,6 @@ interface Props extends TransactionTableProps, WithStyles<typeof styles> {}
 
 class TransactionsTable extends React.Component<Props, TransactionsTableState> {
   public readonly state = {
-    rowsPerPage: 5,
     phoneHook: null,
   };
 
@@ -44,12 +43,6 @@ class TransactionsTable extends React.Component<Props, TransactionsTableState> {
       phoneHook: this.phoneHookRef.current,
     }));
   }
-
-  public readonly rowsChange = (event: React.ChangeEvent<HTMLSelectElement>) => {
-    this.setState({
-      rowsPerPage: Number(event.target.value),
-    });
-  };
 
   public render(): JSX.Element {
     const { classes, onChangeRows } = this.props;
