+++ src/routes/transactions/components/desktop/TransactionsTableHeader.tsx
@@ -4,14 +4,12 @@ import Block from "~/components/layout/Block";
 import Hairline from "~/components/layout/Hairline";
 import Spacer from "~/components/layout/Spacer";
 import Typography from "~/components/layout/Typography";
-import { md } from "~/theme/variables";
 import ColumnHeader from "./ColumnHeader";
 
 const styles = createStyles({
   header: {
     display: "flex",
     alignItems: "center",
-    margin: `${md} 0`,
   },
   title: {
     flex: "1 0 40px",
@@ -22,6 +20,7 @@ interface Props extends WithStyles<typeof styles> {}
 
 const TransactionsTableHeader = ({ classes }: Props) => (
   <React.Fragment>
+    <Block margin="md" />
     <Block padding="lg" className={classes.header}>
       <Typography variant="subtitle2" weight="semibold" className={classes.title}>
         Transactions
@@ -31,6 +30,7 @@ const TransactionsTableHeader = ({ classes }: Props) => (
       <Spacer order={1} />
       <ColumnHeader name="Amount" alignRight />
     </Block>
+    <Block margin="md" />
     <Hairline />
   </React.Fragment>
 );
