+++ src/routes/transactions/components/phone/TransactionsTableHeader.tsx
@@ -5,14 +5,12 @@ import Hairline from "~/components/layout/Hairline";
 import Img from "~/components/layout/Image";
 import Spacer from "~/components/layout/Spacer";
 import Typography from "~/components/layout/Typography";
-import { md } from "~/theme/variables";
 import sorting from "../../assets/sorting.svg";
 
 const styles = createStyles({
   header: {
     display: "flex",
     alignItems: "center",
-    margin: `${md} 0`,
   },
 });
 
@@ -20,6 +18,7 @@ interface Props extends WithStyles<typeof styles> {}
 
 const TransactionsTableHeader = ({ classes }: Props) => (
   <React.Fragment>
+    <Block margin="md" />
     <Block padding="lg" className={classes.header}>
       <Typography variant="subtitle2" weight="semibold">
         Transactions
@@ -27,6 +26,7 @@ const TransactionsTableHeader = ({ classes }: Props) => (
       <Spacer order={1} />
       <Img src={sorting} width={24} height={24} alt="Sorting" />
     </Block>
+    <Block margin="md" />
     <Hairline />
   </React.Fragment>
 );
